Consolidate backend and update docs
