--- requirements.txt
@@ -1,4 +1,0 @@
-Flask
-Flask-Cors
-Werkzeug
-requests
